Create Sum of bit differences.py
Write a program to find the sum of bit differences in all pairs that can be formed from array elements n. Bit difference of a pair (x, y) is a count of different bits at the same positions in binary representations of x and y. For example, a bit difference for 2 and 7 is 2. The binary representation of 2 is 010 and 7 is 111 ( first and last bits differ in two numbers).

We traverse from 0 to 31 and count numbers with ith bit set. Let this count be ‘count’. There would be “n-count” numbers with ith bit not set. So count of differences at ith bit would be “count * (n-count) * 2”, the reason for this formula is as every pair having one element which has set the bit at ith position and a second element having an unset bit at ith position contributes exactly 1 to sum, therefore total permutation count will be count*(n-count) and multiply by 2 is due to one more repetition of all this type of pair